Graphs now draw their lines into RenderTextures for better compositing.
