+++ Prefabs/UI/Graph.prefab
@@ -232,6 +232,18 @@ PrefabInstance:
       insertIndex: -1
       addedObject: {fileID: 4791222170181079484}
   m_SourcePrefab: {fileID: 100100000, guid: 710ce0f3ad8c8d04a9c31eddb197fc49, type: 3}
+--- !u!114 &6763719554259135016 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 5534556083337456922, guid: 710ce0f3ad8c8d04a9c31eddb197fc49,
+    type: 3}
+  m_PrefabInstance: {fileID: 1230395801729287986}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 0
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1344c3c82d62a2a41a3576d8abb8e3ea, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &7372585892862200611 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: 8593872736150834193, guid: 710ce0f3ad8c8d04a9c31eddb197fc49,
@@ -254,6 +266,8 @@ MonoBehaviour:
   - name: Value
     color: {r: 0, g: 1, b: 0, a: 1}
   graphUI: {fileID: 8304128442010129054}
+  linesArea: {fileID: 6763719554259135016}
+  material: {fileID: 2100000, guid: 894cd9d284b82c348b3ad6d99bec7d24, type: 2}
 --- !u!114 &8304128442010129054 stripped
 MonoBehaviour:
   m_CorrespondingSourceObject: {fileID: 7074326688616650156, guid: 710ce0f3ad8c8d04a9c31eddb197fc49,
