--- Scripts/Runtime/UI/GraphLineVisualizer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d4bdf74de8d95864f8448039d352d866
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
